refactor: shared models
