+++ apps/blog/src/api/models/newsletter/dtos.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { email } from "../_shared";
+import { email } from "../shared";
 
 export const newsletterSubscribeDtoSchema = z.object({
   email,
